+++ .env.sample
@@ -4,6 +4,7 @@ SLACK_SIGNING_SECRET="XXXXXXXXXXXXXXXX"
 SLACK_BOT_TOKEN="xoxb-XXXXXXXXXXXXXXXX"
 
 JUDGE_SECRET="secret"
+SUPPORT_SECRET="secret"
 ADMIN_SECRET="something secretive"
 CHALLENGE_URL="https://your-sponsor-challenge.com"
 SLACK_NOTIFICATIONS_WEBHOOK_URL=""
